Adding some stress tests for the task executor.
